Fix unit tests and refactor Alien class due to separete Bomb class
