+++ src/test/java/org/fundacionjala/at15/spaceinvaders/AlienTest.java
@@ -27,13 +27,13 @@ public class AlienTest {
     public void itShouldHaveABullet() {
         Alien theAlien = new Alien(ALIEN_INIT_X, ALIEN_INIT_Y, ALIEN_WIDTH, ALIEN_HEIGHT);
 
-        assertEquals(ALIEN_INIT_X, theAlien.getBullet().getPosX());
-        assertEquals(ALIEN_INIT_Y, theAlien.getBullet().getPosY());
+        assertEquals(ALIEN_INIT_X, theAlien.getBomb().getPosX());
+        assertEquals(ALIEN_INIT_Y, theAlien.getBomb().getPosY());
 
-        Bullet bullet = new Bullet(0, 0, false);
-        theAlien.setBullet(bullet);
+        Bomb bomb = new Bomb(0, 0, false);
+        theAlien.setBomb(bomb);
 
-        assertEquals(0, theAlien.getBullet().getPosX());
-        assertEquals(0, theAlien.getBullet().getPosY());
+        assertEquals(0, theAlien.getBomb().getPosX());
+        assertEquals(0, theAlien.getBomb().getPosY());
     }
 }
